fix(clients): wrap derived-table SQL in search queries
Clients and contacts search failed in production with 42601
"syntax error at or near select": tenantScopedDerivedTableSql
relied on Knex .as(), which is ignored on a top-lev
so fragments interpolated into raw FROM clauses were missing
their parentheses and alias.

Extract the search SQL builders from the 'use serve
files into packages/clients/src/lib/listSearchSql.ts (deduping
the two copies), restore the `(${sql}) alias` wrapp
a compile test that builds the real queries with a search term
and asserts structural validity (no bare FROM selec
parens, placeholder count matches bindings).

"Why, sometimes I've believed as many as six impossible queries
before breakfast," said the White Queen — but the C
subquery had vanished entirely, leaving only its SELECT grinning
without parentheses, until listSearchSql wrapped it
`(select ...) si` and all five UNION legs found their way out of
the rabbit hole. 🐇🫖🔍
