+++ packages/clients/src/actions/queryActionsTenantScoped.contract.test.ts
@@ -3,15 +3,20 @@ import path from 'path';
 import { readFileSync } from 'node:fs';
 
 const source = readFileSync(path.resolve(__dirname, 'queryActions.ts'), 'utf8');
+const searchSqlSource = readFileSync(path.resolve(__dirname, '../lib/listSearchSql.ts'), 'utf8');
 
-function sectionBetween(startMarker: string, endMarker: string): string {
-  const start = source.indexOf(startMarker);
-  const end = source.indexOf(endMarker, start);
+function sectionOf(content: string, startMarker: string, endMarker: string): string {
+  const start = content.indexOf(startMarker);
+  const end = content.indexOf(endMarker, start);
 
   expect(start).toBeGreaterThanOrEqual(0);
   expect(end).toBeGreaterThan(start);
 
-  return source.slice(start, end);
+  return content.slice(start, end);
+}
+
+function sectionBetween(startMarker: string, endMarker: string): string {
+  return sectionOf(source, startMarker, endMarker);
 }
 
 describe('queryActions tenant-scoped query contract', () => {
@@ -41,7 +46,7 @@ describe('queryActions tenant-scoped query contract', () => {
   });
 
   it('uses tenantDb roots and joins for contact list lookups', () => {
-    const contactsByClientSection = sectionBetween('export const getContactsByClient', 'function buildContactListSearchPrefixTsquery');
+    const contactsByClientSection = sectionBetween('export const getContactsByClient', 'export const searchContactListIds');
     const getAllContactsSection = sectionBetween('export const getAllContacts', 'export const findContactByEmailAddress');
     const createOrFindSection = sectionBetween('export const createOrFindContactByEmail', 'function extractNameFromEmail');
 
@@ -65,7 +70,10 @@ describe('queryActions tenant-scoped query contract', () => {
   });
 
   it('uses facade-derived tables for contact list indexed search joins', () => {
-    const searchSection = sectionBetween('export const searchContactListIds', 'export const getAllContacts');
+    const actionSection = sectionBetween('export const searchContactListIds', 'export const getAllContacts');
+    expect(actionSection).toContain('buildContactListSearchQuery(trx, tenant, rawSearch, permissions, user.user_id)');
+
+    const searchSection = sectionOf(searchSqlSource, 'export function buildContactListSearchQuery', 'const bindings: Knex.RawBinding[] = [');
 
     expect(searchSection).toContain('const scopedDb = tenantDb(trx, tenant);');
     expect(searchSection).toContain("const searchIndex = tenantScopedDerivedTableSql(scopedDb, 'app_search_index', 'si');");
